requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
PyYAML>=6.0
tabulate>=0.8.9
colorama>=0.4.4
pymongo>=4.6.0
python-dotenv>=1.0.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
scipy>=1.7.0
plotly>=5.3.0
seaborn>=0.11.0
openpyxl>=3.0.0
supabase>=1.0.3
streamlit>=1.24.0
census>=0.8.19
